move zboost to helper function
