+++ GeneratorInterface/GenFilters/interface/PythiaFilter.h
@@ -31,8 +31,6 @@
 
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 
-#include "HepMC/SimpleVector.h"
-
 //
 // class decleration
 //
@@ -52,9 +50,6 @@ class PythiaFilter : public edm::EDFilter {
 
       virtual bool filter(edm::Event&, const edm::EventSetup&);
    private:
-      // ----------member function -----------------------
-       HepMC::FourVector zboost(const HepMC::FourVector&);
-
       // ----------member data ---------------------------
       
        edm::EDGetTokenT<edm::HepMCProduct> token_;
